Clearified notice on error when output file exists
If for example the meta file cannot be installed, we will die.

Because the downloaded file will be installed before we try to install
the meta file (=when installing the meta file will fail, a new version of
the output file is already installed), the error message we display could
be misinterpreted.

With the new message we hope the to address this problem.