Only replace file when newly downloaded file is different
Previously, when a file was downloaded, for example when no ETags were
available or used and/or Last-Modified header isn't available, we always
replaced the output file with the newly downloaded file making the
--keep=n functionality useless because the same file could replace all
stored versions after n attempts to update the file.

Now, we compare file size and if file size is equal we try to compare
sha256sum or md5sum to determine if the newly downloaded file is different
from the current file.